+++ cmd/mount_test.go
@@ -281,7 +281,7 @@ func (suite *mountTestSuite) TestLibfuseOptions() {
 }
 
 // mount failure test where libfuse options are greater than 8
-func (suite *mountTestSuite) TestLibfuseOptionsMoreThan8() {
+func (suite *mountTestSuite) TestLibfuseIgnoreOptions() {
 	defer suite.cleanupTest()
 
 	mntDir, err := ioutil.TempDir("", "mntdir")
@@ -291,9 +291,9 @@ func (suite *mountTestSuite) TestLibfuseOptionsMoreThan8() {
 	// greater than 8 libfuse options
 	op, err := executeCommandC(rootCmd, "mount", mntDir, fmt.Sprintf("--config-file=%s", confFileMntTest),
 		"-o allow_other", "-o attr_timeout=120", "-o entry_timeout=120", "-o negative_timeout=120",
-		"-o ro", "-o allow_root", "-o default_permissions", "-o umask=755", "-o random_option")
+		"-o ro", "-o allow_root", "-o default_permissions", "-o umask=755", "-o suid,dev")
 	suite.assert.NotNil(err)
-	suite.assert.Contains(op, "invalid FUSE options")
+	suite.assert.Contains(op, "failed to initialize new pipeline")
 }
 
 // mount failure test where a libfuse option is incorrect