+++ component/libfuse/libfuse2_handler_test_wrapper.go
@@ -41,15 +41,16 @@ package libfuse
 import "C"
 import (
 	"errors"
+	"io/fs"
+	"strings"
+	"syscall"
+	"unsafe"
+
 	"github.com/Azure/azure-storage-fuse/v2/common"
 	"github.com/Azure/azure-storage-fuse/v2/common/config"
 	"github.com/Azure/azure-storage-fuse/v2/common/log"
 	"github.com/Azure/azure-storage-fuse/v2/internal"
 	"github.com/Azure/azure-storage-fuse/v2/internal/handlemap"
-	"io/fs"
-	"strings"
-	"syscall"
-	"unsafe"
 
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
@@ -311,9 +312,9 @@ func testOpenAppendFlagDisableWritebackCache(suite *libfuseTestSuite) {
 func testOpenAppendFlagIgnoreAppendFlag(suite *libfuseTestSuite) {
 	defer suite.cleanupTest()
 	suite.cleanupTest() // clean up the default libfuse generated
-	config := "libfuse:\n  ignore-open-flag: true\n"
+	config := "libfuse:\n  ignore-open-flags: true\n"
 	suite.setupTestHelper(config) // setup a new libfuse with a custom config (clean up will occur after the test as usual)
-	suite.assert.True(suite.libfuse.ignoreOpenFlag)
+	suite.assert.True(suite.libfuse.ignoreOpenFlags)
 
 	name := "path"
 	path := C.CString("/" + name)